+++ components/AuthContext/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "AuthContext.js"
+}
