+++ components/List/components/ListItem/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "ListItem.js"
+}
