+++ components/List/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "List.js"
+}
