+++ components/Login/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "Login.js"
+}
