+++ components/Settings/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "Settings.js"
+}
