+++ components/Splash/Splash.js
@@ -1,7 +1,7 @@
 import 'react-native-gesture-handler';
 import React from 'react';
 import {Text, View, Image} from 'react-native';
-import {styles} from '../styles/styles';
+import {styles} from '../../styles/styles';
 
 export default function Splash() {
   return (
