+++ components/Splash/package.json
@@ -1,0 +1,3 @@
+{
+  "main": "Splash.js"
+}
